Refactor script structure, improve user prompts for better clarity, added right click context menu fix, check if it already exists, remove if it is, critical action confirming dialogs

+++ essentials.ps1
@@ -1,3 +1,5 @@
+# https://github.com/kostikasz
+
 # Check for admin rights
 $isAdmin = ([Security.Principal.WindowsPrincipal] `
     [Security.Principal.WindowsIdentity]::GetCurrent()
@@ -6,6 +8,7 @@ $isAdmin = ([Security.Principal.WindowsPrincipal] `
 
 if (-not $isAdmin) {
     Write-Host "Please run this script as Administrator." -ForegroundColor Red
+    Start-Sleep 1
     exit
 }
 
@@ -14,9 +17,7 @@ if (-not $isAdmin) {
 # ALL THE OPTION MENUS
 
 function StartingMenu {
-    Write-Host "
-
-kostikasz IT support quick kit
+    Write-Host "kostikasz IT support quick kit
 
 Choose an option:
 [1] Install essential apps
@@ -28,8 +29,7 @@ Choose an option:
 }
 
 function InstallMenu {
-    Write-Host "
-kostikasz IT support quick kit
+    Write-Host "kostikasz IT support quick kit
 
 ESSENTIAL APPS INSTALLATION
 Choose an option:
@@ -44,8 +44,7 @@ Choose an option:
 }
 
 function InstallMenuBrowsers {
-       Write-Host "
-kostikasz IT support quick kit
+       Write-Host "kostikasz IT support quick kit
 
 BROWSERS INSTALLATION
 Choose an option:
@@ -59,8 +58,7 @@ Choose an option:
 }
 
 function FixMenu {
-       Write-Host "
-kostikasz IT support quick kit
+       Write-Host "kostikasz IT support quick kit
 
 FIX LIST:
 Choose an option:
@@ -73,12 +71,34 @@ Choose an option:
 }
 
 
+function Confirm {
+
+    while ($true) {
+        Clear-Host
+        Write-Host "kostikasz IT support quick kit
+
+Are you sure? Make sure you don't have unsaved work.:
+Choose an option:
+[Y] Y (Yes)     [N] N (No)
+
+       "
+        $choice = Read-Host 'Enter choice'
+
+        switch ($choice.ToLower()) {
+            'y' { return $true }
+            'n' { return $false }
+            default {
+                Write-Host "Invalid option."
+                Start-Sleep 1
+            }
+        }
+    }
+}
 
 
 function InstallingBrowsers {
     while ($true) {
         #Clear-Host
-        Write-Host $option
         $option = InstallMenuBrowsers
 
         if ($option -eq "") {
@@ -90,24 +110,28 @@ function InstallingBrowsers {
 
         switch ($option.ToLower()) {
         "1" {
+            Clear-Host
             Write-Host "Installing Google Chrome"
             winget.exe install -e --id Google.Chrome --source=winget
             Write-Host "Succesfully installed Google Chrome, returning to essential app install menu" -ForegroundColor Green
             return
         }
         "2" {
+            Clear-Host
             Write-Host "Installing Mozilla Firefox"
             winget.exe install -e --id Mozilla.Firefox --source=winget
             Write-Host "Succesfully installed Mozilla Firefox, returning to essential app install menu" -ForegroundColor Green
             return
         }
         "3" {
+            Clear-Host
             Write-Host "Installing Brave"
             winget.exe install -e --id Brave.Brave --source=winget
             Write-Host "Succesfully installed Brave, returning to essential app install menu" -ForegroundColor Green
             return
         }
         "q" {
+            Clear-Host
             Write-Host "Returning to main menu"
             return
         }
@@ -122,8 +146,6 @@ function InstallingBrowsers {
 
 function InstallingEssentials {
     while ($true) {
-        #Clear-Host
-        Write-Host $option
         $option = InstallMenu
 
         if ($option -eq "") {
@@ -139,10 +161,12 @@ function InstallingEssentials {
             InstallingBrowsers
         }
         "2" {
-            Write-Host "Coming soon"
-            return
+            Clear-Host
+            Write-Host "Coming soon" -ForegroundColor Yellow
+            continue
         }
         "q" {
+            Clear-Host
             Write-Host "Returning to main menu"
             return
         }
@@ -154,10 +178,33 @@ function InstallingEssentials {
     }
 }
 
+function RevertChanges {
+
+    while ($true) {
+        <# Clear-Host #>
+        Write-Host "kostikasz IT support quick kit
+
+Want to revert changes? Make sure you don't have unsaved work:
+Choose an option:
+[Y] Y (Yes)     [Y] N (No)
+
+       "
+        $choice = Read-Host 'Enter choice'
+
+        switch ($choice.ToLower()) {
+            'y' { return $true }
+            'n' { return $false }
+            default {
+                Write-Host "Invalid option."
+                Start-Sleep 1
+            }
+        }
+    }
+}
+
+
 function ApplyFixes {
     while ($true) {
-        #Clear-Host
-        Write-Host $option
         $option = FixMenu
 
         if ($option -eq "") {
@@ -169,21 +216,58 @@ function ApplyFixes {
 
         switch ($option.ToLower()) {
         "1" {
-            Write-Host "Applying right-click context menu fix"
-            reg.exe add "HKCU\Software\Classes\CLSID\{86ca1aa0-34aa-4e8b-a509-50c905bae2a2}\InprocServer32" /f /ve
-            Stop-Process -ProcessName explorer -Force #TODO make silent
-            Write-Host "Fix applied succesfully, returning to menu"
+            <# Clear-Host #>
+            switch (Confirm) {
+                $true {
+                    $fixstatus=reg.exe query "HKCU\Software\Classes\CLSID\{86ca1aa0-34aa-4e8b-a509-50c905bae2a2}\InprocServer32"
+                    if ($fixstatus)  {
+                        Write-Host "Error. The fix is already applied." -ForegroundColor Red
+                        if (RevertChanges) {
+                            Write-Host "Reverting right-click context menu fix"
+                            reg.exe delete "HKCU\Software\Classes\CLSID\{86ca1aa0-34aa-4e8b-a509-50c905bae2a2}\InprocServer32" /f /ve
+                            Stop-Process -ProcessName explorer -Force #TODO make silent
+                            Clear-Host
+                            Write-Host "Fix deleted succesfully, returning to menu"
+                            continue
+                        } else {
+                            Clear-Host
+                            Write-Host "Didn't make any changes. Returning to menu"
+                            continue
+                        }
+
+                    } else {
+
+                        #Make it check if the registery edit already exists
+                        Write-Host "Applying right-click context menu fix"
+                        reg.exe add "HKCU\Software\Classes\CLSID\{86ca1aa0-34aa-4e8b-a509-50c905bae2a2}\InprocServer32" /f /ve
+                        Stop-Process -ProcessName explorer -Force #TODO make silent
+                        Clear-Host
+                        Write-Host "Fix applied succesfully, returning to menu"
+                        continue
+                    }
+                    
+
+                }
+                default {
+                    Clear-Host
+                    Write-Host "Cancelled. Returning to fix menu"
+                    continue
+                } 
+            }
         }
         "2" {
-            Write-Host "Coming soon"
-            return
+            Clear-Host
+            Write-Host "Coming soon" -ForegroundColor Yellow
+            continue
         }
         "q" {
+            Clear-Host
             Write-Host "Returning to main menu"
             return
         }
         default {
             Write-Host "Invalid option."
+            continue
         }
        }
        
@@ -198,8 +282,7 @@ function ApplyFixes {
 # Main loop of the starting menu
 
 while ($true) {
-    <# Clear-Host #>
-
+    Clear-Host
     $option = ""
     $option = StartingMenu
     if ($option -eq "") {
@@ -210,10 +293,12 @@ while ($true) {
 
     switch ($option.ToLower()) {
         "1" {
+                Clear-Host
                 Write-Host "Opening essential app installation menu"
                 InstallingEssentials
         }
         "2" {
+                Clear-Host
                 Write-Host "Opening fix menu"
                 ApplyFixes
         }
@@ -223,6 +308,7 @@ while ($true) {
         }
             default {
             Write-Host "Invalid option."
+            return
         }
     }
 }
